script/requirements_category_crawler.txt
# Yahoo!ニュース分类新闻爬虫依赖包
# 基础依赖
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=0.19.0

# LLM集成
openai>=0.27.0

# 可选依赖（用于更好的性能）
lxml>=4.9.0
html5lib>=1.1

# 开发依赖（可选）
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
